Add mock user data functions
